+++ ruby/CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.9.12
 * Add 22 optional parameters to sling post agents support #22
+* Add AEM Health Check Servlet config support
 
 ### 0.9.11
 * Rename AEM Password Reset config property to pwdreset_authorizables